requirements.txt
adblockparser==0.7
attrs==23.1.0
beautifulsoup4==4.12.2
browsermob-proxy==0.8.0
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
DateTime==5.1
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
numpy==1.25.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psutil==5.9.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyVirtualDisplay==3.0
requests==2.31.0
selenium==4.16.0
Selenium-Screenshot==2.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tld==0.13
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
tzdata==2023.3
undetected-chromedriver==3.5.0
urllib3==2.0.3
webdriver-manager==4.0.1
websockets==11.0.3
wsproto==1.2.0
zope.interface==6.0
